- renamed files

+++ BlogAPI/services/Impl/AuthService.cs
@@ -1,5 +1,4 @@
-﻿using System.Security.Claims;
-using BlogAPI.Data;
+﻿using BlogAPI.Data;
 using BlogAPI.Entities;
 using BlogAPI.Exceptions;
 using BlogAPI.Models.Request;
@@ -14,20 +13,18 @@ public class AuthService : IAuthService
 {
     private readonly ApplicationDbContext _db;
     private readonly IJwtService _jwtService;
-    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public AuthService(ApplicationDbContext db, IJwtService jwtService, IHttpContextAccessor httpContextAccessor)
+    public AuthService(ApplicationDbContext db, IJwtService jwtService)
     {
         _db = db;
         _jwtService = jwtService;
-        _httpContextAccessor = httpContextAccessor;
     }
 
 
-    public async Task<RegistrationResponse> Register(User user)
+    public async Task<RegistrationResponse> Register(RegisterRequest registerRequest)
     {
-        await IsUserUnique(user.Email);
-        var newUser = CreateHashUser(user);
+        await IsUserUnique(registerRequest.Email);
+        var newUser = CreateHashUser(registerRequest);
         await _db.Users.AddAsync(newUser);
         await _db.SaveChangesAsync();
         return new RegistrationResponse { Email = newUser.Email, FullName = newUser.FullName };
@@ -38,11 +35,11 @@ public class AuthService : IAuthService
     {
         var user = await GetUserByEmailAsync(loginRequest.Email);
         CheckIsValidPassword(loginRequest.Password, user.Password);
-        var existingRefreshToken = await _jwtService.GetRefreshTokenByEmailAsync(loginRequest.Email);
+        var existingRefreshToken = await _jwtService.GetRefreshTokenByGuidAsync(user.Id);
         var accessToken = _jwtService.GenerateAccessToken(user);
         
         var refreshToken = existingRefreshToken ?? _jwtService.GenerateRefreshToken(user);
-        if (refreshToken != existingRefreshToken) await _jwtService.SaveRefreshTokenAsync(refreshToken, loginRequest.Email);
+        if (refreshToken != existingRefreshToken) await _jwtService.SaveRefreshTokenAsync(refreshToken, user.Id);
         
         return new LoginResponse { AccessToken = accessToken, RefreshToken = refreshToken };
     }
@@ -50,26 +47,27 @@ public class AuthService : IAuthService
 
     public async Task<IActionResult> Logout()
     {
-        var userEmail = _httpContextAccessor.HttpContext?.User.FindFirstValue(ClaimTypes.Email);
-        var refreshToken = await _jwtService.GetRefreshTokenByEmailAsync(userEmail);
+        var userId = await _jwtService.GetUserGuidFromTokenAsync();
+        if (userId == Guid.Empty) throw new UserNotFoundException("User not found");
+        var refreshToken = await _jwtService.GetRefreshTokenByGuidAsync(userId);
         if (refreshToken == null) throw new NullTokenException("Refresh token not found");
         await _jwtService.RemoveRefreshTokenAsync(refreshToken);
         
         return new OkResult();
     }
 
-    public async Task<RefreshResponse> Refresh(RefreshRequest refreshRequest)
+    public RefreshResponse Refresh(RefreshRequest refreshRequest)
     {
-        var userEmail = _jwtService.GetEmailFromRefreshToken(refreshRequest.RefreshToken);
-        await _jwtService.ValidateRefreshTokenAsync(refreshRequest.RefreshToken);
-        var user = GetUserByEmail(userEmail);
+        var userGuid = _jwtService.GetGuidFromRefreshToken(refreshRequest.RefreshToken);
+        _jwtService.ValidateRefreshToken(refreshRequest.RefreshToken);
+        var user = GetUserByGuid(userGuid);
         var accessToken = _jwtService.GenerateAccessToken(user);
         return new RefreshResponse { AccessToken = accessToken };
     }
 
     
 
-    public async Task IsUserUnique(string? userEmail)
+    private async Task IsUserUnique(string? userEmail)
     {
         var user = await _db.Users.FirstOrDefaultAsync(u => u.Email == userEmail);
         if (user != null) throw new UserAlreadyExistsException("User already exists");
@@ -90,27 +88,26 @@ public class AuthService : IAuthService
         if (user == null) throw new UserNotFoundException("Wrong email or password");
         return user;
     }
-    
-    private User GetUserByEmail(string loginRequestEmail)
+    private User GetUserByGuid(Guid userId)
     {
-        var user = _db.Users.FirstOrDefault(u => u.Email == loginRequestEmail);
-        if (user == null) throw new UserNotFoundException("Wrong email or password");
+        var user =  _db.Users.FirstOrDefault(u => u.Id == userId);
+        if (user == null) throw new UserNotFoundException("User not found");
         return user;
     }
 
 
-    private static User CreateHashUser(User user)
+    private static User CreateHashUser(RegisterRequest registerRequest)
     {
-        var passwordHash = BCrypt.Net.BCrypt.HashPassword(user.Password);
+        var passwordHash = BCrypt.Net.BCrypt.HashPassword(registerRequest.Password);
         return new User
         {
-            FullName = user.FullName,
-            Phone = user.Phone,
+            FullName = registerRequest.FullName,
+            Phone = registerRequest.PhoneNumber,
             Password = passwordHash,
-            BirthDate = user.BirthDate.ToUniversalTime(),
-            Email = user.Email,
-            Gender = user.Gender,
-            Id = user.Id,
+            BirthDate = registerRequest.BirthDate.ToUniversalTime(),
+            Email = registerRequest.Email,
+            Gender = registerRequest.Gender,
+            Id = Guid.NewGuid(),
             CreatedAt = DateTime.UtcNow.ToUniversalTime()
         };
     }
